Limited Material for MkDocs version range
The release of Material for MkDocs 6.2 is about to happen, which might contain some breaking changes due to the modifications we made on the ToIP template. Essentially, we can remove some of the modifications, as those things are now supported by the upstream repository. This is just a safety measure.